// src/api-client/models/chat-message-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ParticipantInfo } from './participant-info';

/**
 * 
 * @export
 * @interface ChatMessageResponse
 */
export interface ChatMessageResponse {
    /**
     * 
     * @type {string}
     * @memberof ChatMessageResponse
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatMessageResponse
     */
    'conversationId'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ChatMessageResponse
     */
    'me'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ChatMessageResponse
     */
    'message'?: string;
    /**
     * 
     * @type {ParticipantInfo}
     * @memberof ChatMessageResponse
     */
    'sender'?: ParticipantInfo;
    /**
     * 
     * @type {string}
     * @memberof ChatMessageResponse
     */
    'createdDate'?: string;
}